+++ src/nvim/ugrid.c
@@ -73,7 +73,7 @@ void ugrid_scroll(UGrid *grid, int top, int bot, int left, int right, int count)
     UCell *target_row = grid->cells[i] + left;
     UCell *source_row = grid->cells[i + count] + left;
     memcpy(target_row, source_row,
-           sizeof(UCell) * (size_t)(right - left + 1));
+           sizeof(UCell) * (size_t)right - (size_t)left + 1);
   }
 }
